Урок 3. Домашняя работа.

+++ app/src/main/java/ru/zfix27r/movies/domain/model/TopResModel.kt
@@ -1,7 +1,7 @@
 package ru.zfix27r.movies.domain.model
 
 data class TopResModel(
-    val id: Long,
+    val id: Int,
     val nameRu: String,
     val posterUrlPreview: String
 )
